Move preview logic into app.loop
